+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## v1.0.2
+
+* Fixed migration files
+
+## v1.0.1
+
+* Fixed broken install generator
+
 ## v1.0.1
 
 * Initial Release